// common/enums.ts
// lead page mode type
export enum LEAD_MODE_TYPE_ENUM {
    ADD = "add",
    VIEW = "view",
    EDIT = "edit"
};

// leads status enum
export enum LEADS_STATUS_ENUM {
    NEW = "new",
    CONTACTED = "contacted",
    QUALIFIED = "qualified",
    LOST = "lost"
};

// toast type
export enum TOAST_TYPE_ENUM {
    SUCCESS = "success",
    ERROR = "error"
};